# Build: Update CircleCI configuration to remove build-sandboxes job