--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -396,7 +396,7 @@ jobs:
       parallelism:
         type: integer
     executor:
-      class: medium
+      class: large
       name: sb_node_22_browsers
     parallelism: << parameters.parallelism >>
     steps:
@@ -408,8 +408,24 @@ jobs:
           name: Enable Corepack
           command: sudo corepack enable yarn
       - run:
+          name: Starting Event Collector
+          command: yarn jiti ./event-log-collector.ts
+          working_directory: scripts
+          background: true
+      - run:
           name: Creating Sandboxes
-          command: yarn task --task sandbox --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox) --no-link --start-from=never --junit
+          command: yarn task --task sandbox --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox) --no-link --start-from=build --junit
+          environment:
+            STORYBOOK_TELEMETRY_URL: "http://localhost:6007/event-log"
+      - run:
+          name: Verifying Telemetry
+          command: |
+            TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox)
+            if [[ $TEMPLATE != bench/* ]]
+            then
+              yarn jiti ./event-log-checker.ts build $TEMPLATE
+            fi
+          working_directory: scripts
       - report-workflow-on-failure:
           template: $(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox)
       - persist_to_workspace:
@@ -438,47 +454,6 @@ jobs:
           template: $(yarn get-template --cadence << pipeline.parameters.workflow >> --task smoke-test)
       - store_test_results:
           path: test-results
-  build-sandboxes:
-    parameters:
-      parallelism:
-        type: integer
-    executor:
-      class: large
-      name: sb_node_22_browsers
-    parallelism: << parameters.parallelism >>
-    steps:
-      - git-shallow-clone/checkout_advanced:
-          clone_options: "--depth 1 --verbose"
-      - attach_workspace:
-          at: .
-      - run:
-          name: Starting Event Collector
-          command: yarn jiti ./event-log-collector.ts
-          working_directory: scripts
-          background: true
-      - run:
-          name: Building Sandboxes
-          command: yarn task --task build --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task build) --no-link --start-from=never --junit
-          environment:
-            STORYBOOK_TELEMETRY_URL: "http://localhost:6007/event-log"
-      - run:
-          name: Verifying Telemetry
-          command: |
-            TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task build)
-            if [[ $TEMPLATE != bench/* ]]
-            then
-              yarn jiti ./event-log-checker.ts build $TEMPLATE
-            fi
-          working_directory: scripts
-      - report-workflow-on-failure:
-          template: $(yarn get-template --cadence << pipeline.parameters.workflow >> --task build)
-      - store_test_results:
-          path: test-results
-      - persist_to_workspace:
-          root: .
-          paths:
-            - sandbox/*/bench/*.json
-            - sandbox/*/storybook-static
   test-runner-production:
     parameters:
       parallelism:
@@ -837,18 +812,14 @@ workflows:
           parallelism: 14
           requires:
             - build
-      - build-sandboxes:
-          parallelism: 14
-          requires:
-            - create-sandboxes
       - chromatic-sandboxes:
           parallelism: 11
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - e2e-production:
           parallelism: 9
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - e2e-dev:
           parallelism: 1
           requires:
@@ -856,7 +827,7 @@ workflows:
       - test-runner-production:
           parallelism: 9
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - vitest-integration:
           parallelism: 5
           requires:
@@ -864,7 +835,7 @@ workflows:
       - bench-sandboxes:
           parallelism: 5
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - test-ui-testing-module:
           requires:
             - build
@@ -913,18 +884,14 @@ workflows:
           parallelism: 20
           requires:
             - build
-      - build-sandboxes:
-          parallelism: 20
-          requires:
-            - create-sandboxes
       - chromatic-sandboxes:
           parallelism: 17
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - e2e-production:
           parallelism: 15
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - e2e-dev:
           parallelism: 1
           requires:
@@ -932,7 +899,7 @@ workflows:
       - test-runner-production:
           parallelism: 15
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - vitest-integration:
           parallelism: 5
           requires:
@@ -949,7 +916,7 @@ workflows:
       - bench-sandboxes:
           parallelism: 5
           requires:
-            - build-sandboxes
+            - create-sandboxes
       # TODO: reenable once we find out the source of flakyness
       # - test-runner-dev:
       #     parallelism: 4
@@ -990,18 +957,14 @@ workflows:
       # - smoke-test-sandboxes: # disabled for now
       #     requires:
       #       - create-sandboxes
-      - build-sandboxes:
-          parallelism: 37
-          requires:
-            - create-sandboxes
       - chromatic-sandboxes:
           parallelism: 34
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - e2e-production:
           parallelism: 32
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - e2e-dev:
           parallelism: 1
           requires:
@@ -1009,7 +972,7 @@ workflows:
       - test-runner-production:
           parallelism: 32
           requires:
-            - build-sandboxes
+            - create-sandboxes
       - vitest-integration:
           parallelism: 11
           requires:
@@ -1044,7 +1007,7 @@ workflows:
       - bench-sandboxes:
           parallelism: 5
           requires:
-            - build-sandboxes
+            - create-sandboxes
 
       # TODO: reenable once we find out the source of flakyness
       # - test-runner-dev: