--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -408,18 +408,16 @@ jobs:
           name: Enable Corepack
           command: sudo corepack enable yarn
       - run:
-          name: Create Sandboxes
-          command: yarn task --task sandbox --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox) --no-link --start-from=never --junit
-          environment:
-            STORYBOOK_TELEMETRY_URL: "http://localhost:6007/event-log"
-      - run:
           name: Starting Event Collector
           command: yarn jiti ./event-log-collector.ts
           working_directory: scripts
           background: true
       - run:
-          name: Build Sandboxes
-          command: yarn task --task build --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task build) --no-link --start-from=never --junit
+          name: Create Sandboxes
+          command: |
+            TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox)
+            yarn task --task sandbox --template "${TEMPLATE}" --no-link --start-from=never --junit
+            yarn task --task build --template "${TEMPLATE}" --no-link --start-from=never --junit
           environment:
             STORYBOOK_TELEMETRY_URL: "http://localhost:6007/event-log"
       - run: