--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -418,17 +418,12 @@ jobs:
             TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox)
             yarn task --task sandbox --template "${TEMPLATE}" --no-link --start-from=never --junit
             yarn task --task build --template "${TEMPLATE}" --no-link --start-from=never --junit
-          environment:
-            STORYBOOK_TELEMETRY_URL: "http://localhost:6007/event-log"
-      - run:
-          name: Verifying Telemetry
-          command: |
-            TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox)
             if [[ $TEMPLATE != bench/* ]]
             then
-              yarn jiti ./event-log-checker.ts build $TEMPLATE
+              yarn jiti ./event-log-checker.ts build $TEMPLATE --cwd scripts
             fi
-          working_directory: scripts
+          environment:
+            STORYBOOK_TELEMETRY_URL: "http://localhost:6007/event-log"
       - report-workflow-on-failure:
           template: $(yarn get-template --cadence << pipeline.parameters.workflow >> --task sandbox)
       - persist_to_workspace: