--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -405,9 +405,6 @@ jobs:
       - attach_workspace:
           at: .
       - run:
-          name: Enable Corepack
-          command: corepack enable --install-directory="/home/circleci/bin"
-      - run:
           name: Starting Event Collector
           command: yarn jiti ./event-log-collector.ts
           working_directory: scripts