--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -405,13 +405,13 @@ jobs:
       - attach_workspace:
           at: .
       - run:
+          name: Enable Corepack
+          command: sudo corepack enable yarn
+      - run:
           name: Starting Event Collector
           command: yarn jiti ./event-log-collector.ts
           working_directory: scripts
           background: true
-      - run:
-          name: Enable Corepack
-          command: sudo corepack enable yarn
       - run:
           name: Create Sandboxes
           command: |