--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -211,7 +211,7 @@ jobs:
       - cancel-workflow-on-failure
   bench-packages:
     executor:
-      class: medium
+      class: small
       name: sb_node_22_classic
     steps:
       - git-shallow-clone/checkout_advanced:
@@ -268,24 +268,18 @@ jobs:
       - nx/set-shas:
           main-branch-name: "next"
           workflow-name: << pipeline.parameters.workflow >>
-      - run:
-          name: install in scripts
-          command: |
-            cd scripts
-            yarn install
-      - run:
-          name: install in code
-          command: |
-            cd code
-            yarn install
+     - restore_cache:
+          name: Restore Yarn cache
+          keys:
+            - build-yarn-2-cache-v4--{{ checksum "code/yarn.lock" }}--{{ checksum "scripts/yarn.lock" }}
       - run:
           name: Compile
           command: |
-            yarn task --task compile --start-from=compile --no-link --debug
+            yarn task --task compile --start-from=auto --no-link --debug
       - run:
           name: Check
           command: |
-            yarn task --task check --start-from=check --no-link --debug
+            yarn task --task check --start-from=auto --no-link --debug
       - run:
           name: Ensure no changes pending
           command: |
@@ -429,6 +423,7 @@ jobs:
           root: .
           paths:
             - sandbox
+            - !sandbox/**/node_modules
       - store_test_results:
           path: test-results
   smoke-test-sandboxes:
@@ -464,6 +459,10 @@ jobs:
           clone_options: "--depth 1 --verbose"
       - attach_workspace:
           at: .
+      - run: Install sandbox dependencies
+        command: |
+          TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task test-runner)
+          cd sandbox/$TEMPLATE && yarn
       - run:
           name: Running Test Runner
           command: yarn task --task test-runner --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task test-runner) --no-link --start-from=never --junit
@@ -484,6 +483,10 @@ jobs:
           clone_options: "--depth 1 --verbose"
       - attach_workspace:
           at: .
+      - run: Install sandbox dependencies
+        command: |
+          TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task vitest-integration)
+          cd sandbox/$TEMPLATE && yarn
       - run:
           name: Running story tests in Vitest
           command: yarn task --task vitest-integration --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task vitest-integration) --no-link --start-from=never --junit
@@ -523,6 +526,10 @@ jobs:
       - checkout
       - attach_workspace:
           at: .
+      - run: Install sandbox dependencies
+        command: |
+          TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task chromatic)
+          cd sandbox/$TEMPLATE && yarn
       - run:
           name: Running Chromatic
           command: yarn task --task chromatic --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task chromatic) --no-link --start-from=never --junit
@@ -543,6 +550,10 @@ jobs:
           clone_options: "--depth 1 --verbose"
       - attach_workspace:
           at: .
+      - run: Install sandbox dependencies
+        command: |
+          TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task e2e-tests)
+          cd sandbox/$TEMPLATE && yarn
       - run:
           name: Running E2E Tests
           command: yarn task --task e2e-tests --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task e2e-tests) --no-link --start-from=never --junit
@@ -566,6 +577,10 @@ jobs:
           clone_options: "--depth 1 --verbose"
       - attach_workspace:
           at: .
+      - run: Install sandbox dependencies
+        command: |
+          TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task e2e-tests-dev)
+          cd sandbox/$TEMPLATE && yarn
       - run:
           name: Running E2E Tests
           command: yarn task --task e2e-tests-dev --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task e2e-tests-dev) --no-link --start-from=never --junit
@@ -589,6 +604,10 @@ jobs:
           clone_options: "--depth 1 --verbose"
       - attach_workspace:
           at: .
+      - run: Install sandbox dependencies
+        command: |
+          TEMPLATE=$(yarn get-template --cadence << pipeline.parameters.workflow >> --task bench)
+          cd sandbox/$TEMPLATE && yarn
       - run:
           name: Running Bench
           command: yarn task --task bench --template $(yarn get-template --cadence << pipeline.parameters.workflow >> --task bench) --no-link --start-from=never --junit