--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -415,7 +415,7 @@ jobs:
             if [[ $TEMPLATE != bench/* ]]; then
               yarn --cwd scripts jiti ./event-log-checker.ts build $TEMPLATE
             fi
-            cd sandbox/$TEMPLATE && rm -rf node_modules
+            cd sandbox/$(yarn get-sandbox-dir --template $TEMPLATE) && rm -rf node_modules
           environment:
             STORYBOOK_TELEMETRY_URL: "http://localhost:6007/event-log"
       - report-workflow-on-failure: