Closes #

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

- Merged build-sandboxes into create-sandboxes (saved 3-5 mins)
- Rebalanced CI resources so that jobs run more efficiently. Non-blocking tasks can run longer if we can save some resources. Blocking tasks should run faster.
- Made the workspace persistence more efficient: The build-sandboxes step persisted all sandboxes and their node_modules leading to a huge persisted workspace. Re-attaching it in a follow-up step took up to 2:30 mins. Now node_modules from sandboxes do not persist anymore, decreasing the “Attach workspace” step of the last jobs by another 1:30 mins.
- Updated yarn from 4.3.0 to 4.6.0

Workflow `daily` before:
![image (5)](https://github.com/user-attachments/assets/70a6fb45-6727-40dd-90b5-f2874ebb14cb)

Workflow `daily` after:
![image (6)](https://github.com/user-attachments/assets/41907d32-eb4a-4d59-8faf-3e8b65d59687)


## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  77.9 MB | 77.9 MB | 0 B | **1.54** | 0% |
| initSize |  131 MB | 131 MB | 614 B | 1.12 | 0% |
| diffSize |  53 MB | 53 MB | 614 B | -0.26 | 0% |
| buildSize |  7.17 MB | 7.17 MB | 0 B | **-1.36** | 0% |
| buildSbAddonsSize |  1.85 MB | 1.85 MB | 0 B | **1.36** | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.86 MB | 1.86 MB | 0 B | **-1.36** | 0% |
| buildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.91 MB | 3.91 MB | 0 B | **-1.36** | 0% |
| buildPreviewSize |  3.26 MB | 3.26 MB | 0 B | **-1.36** | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  24.7s | 24.8s | 107ms | **1.34** | 0.4% |
| generateTime |  20.2s | 21.7s | 1.5s | 0.54 | 7% |
| initTime |  13.7s | 12.9s | -788ms | -0.89 | -6.1% |
| buildTime |  9.9s | 8.9s | -1s -25ms | -0.55 | -11.5% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  4.9s | 4.5s | -329ms | -0.88 | -7.2% |
| devManagerResponsive |  3.5s | 3.4s | -142ms | -0.7 | -4.2% |
| devManagerHeaderVisible |  621ms | 640ms | 19ms | 0.46 | 3% |
| devManagerIndexVisible |  651ms | 649ms | -2ms | 0.05 | -0.3% |
| devStoryVisibleUncached |  2s | 3.5s | 1.5s | **6.24** | 🔺42.6% |
| devStoryVisible |  652ms | 667ms | 15ms | 0.25 | 2.2% |
| devAutodocsVisible |  446ms | 626ms | 180ms | 0.88 | 28.8% |
| devMDXVisible |  514ms | 618ms | 104ms | 1.23 | 16.8% |
| buildManagerHeaderVisible |  605ms | 636ms | 31ms | 0.37 | 4.9% |
| buildManagerIndexVisible |  709ms | 717ms | 8ms | 0.26 | 1.1% |
| buildStoryVisible |  594ms | 581ms | -13ms | -0.05 | -2.2% |
| buildAutodocsVisible |  437ms | 478ms | 41ms | -0.28 | 8.6% |
| buildMDXVisible |  458ms | 543ms | 85ms | 0.8 | 15.7% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Here's my summary of the key changes in this PR:

This PR optimizes CircleCI configuration by merging jobs and improving workspace persistence, while also updating core dependencies.

- Merges `build-sandboxes` into `create-sandboxes` job, saving 3-5 minutes per workflow
- Optimizes workspace persistence by excluding sandbox `node_modules`, reducing attach time by ~1:30 minutes
- Updates Node.js from 22.6.0 to 22.13.1 and Yarn from 4.3.0 to 4.6.0 across all configuration files
- Rebalances CircleCI resource classes (xlarge -> large, large -> medium+) for better efficiency
- Adds new `get-sandbox-dir.ts` script to handle sandbox directory name generation

The changes show improved build times but introduce some concerning performance regressions in dev environment responsiveness that should be investigated.



<!-- /greptile_comment -->